requirements.txt
pandas
kagglehub
python-dotenv
sqlalchemy
psycopg2-binary
apache-airflow
